refactor(typings): renamings
