+++ src/typing/array.ts
