+++ src/typing/colors.ts
