+++ src/typing/digit.ts
