+++ src/typing/index.ts
@@ -1,0 +1,3 @@
+export type { FixedArray } from './array'
+export type { Digit } from './digit'
+export type { HTTPCode } from './http'
